Updated map connections
Reinstated connections between maps. Removed connections from other, non-connected maps.

Bug report: tiles unique to the tilesets of neighbouring maps do not load in Porymap in the "connections" view. This was not the case in the previous project, and may be due to faulty manual porting (i.e. I missed something...).